Reworked all stat services and repositories

+++ backend/src/main/java/com/gw/backend/models/stats/Statistics.java
@@ -1,4 +1,4 @@
 package com.gw.backend.models.stats;
 
-public record Statistics(int likes, int total, int percentage) {
+public record Statistics(Long likes, Long total, int percentage) {
 }
